Deduplicate rspec_custom_matchers to just one copy
