+++ stash-harvester/stash-harvester.gemspec
@@ -39,6 +39,7 @@ Gem::Specification.new do |s| # rubocop:disable Metrics/BlockLength
   s.add_dependency 'stash-wrapper'
 
   s.add_development_dependency 'bundler', '~> 1.7'
+  s.add_development_dependency 'diffy', '~> 3.1'
   s.add_development_dependency 'equivalent-xml', '~> 0.6.0'
   s.add_development_dependency 'github-markup', '~> 1.4'
   s.add_development_dependency 'nokogiri', '~> 1.8'
