+++ stash-sword/stash-sword.gemspec
@@ -32,6 +32,7 @@ Gem::Specification.new do |s| # rubocop:disable Metrics/BlockLength
   s.add_development_dependency 'equivalent-xml', '~> 0.6.0'
 
   s.add_development_dependency 'bundler', '~> 1.7'
+  s.add_development_dependency 'diffy', '~> 3.1'
   s.add_development_dependency 'nokogiri', '~> 1.8'
   s.add_development_dependency 'rake', '~> 12.0'
   s.add_development_dependency 'rspec', '~> 3.3'
